requirements.txt
mlflow>=2.14.1 
openai==1.52.1
transformers 
torch==2.5.0
torchvision==0.20.0
evaluate==0.4.3
datasets 
tiktoken 
fastapi 
rouge_score 
textstat 
tenacity 
plotly 
ipykernel 
nbformat>=5.10.4
python-dotenv==1.0.1

#Rag système
chromadb==0.4.15
langchain==0.0.348
fastapi
# pip install 'mlflow[genai]'
# pip install databricks-sdk --upgrade

duckduckgo-search==6.3.2
beautifulsoup4